HZ-SAGE/backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-docx
fpdf2
pandas
numpy
langchain
langchain-core
langchain-openai
pydantic
python-dotenv
